Implement container component approach for add & edit user
